+++ src/Vortice.Direct3D12/Constants.cs
@@ -511,7 +511,6 @@ namespace Vortice.Direct3D12
         public const int WHQLDrawIndexedIndexCount2ToExp = 25;
         public const int WHQLDrawVertexCount2ToExp = 25;
 
-        public const int PIX_EVENT_UNICODE_VERSION = 0;
-        public const int PIX_EVENT_ANSI_VERSION = 1;
+        internal const int PIX_EVENT_UNICODE_VERSION = 0;
     }
 }