+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Release:     1.8.XX (October 2020)
 KEY CHANGES:
 - [ENH] DirectComposition: Initial DirectComposition support.
 - [ENH] DXGI: Initial WinUI headers generation under WinUI folder + namespace.
+- [FIX] D3D12: BeginEvent, SetMarker for ID3D12CommandQueue and ID3D12GraphicsCommandList.
 
 -----------------------------------------------
 Release:     1.8.17 (October 2020)